implement new info modal for location facts
