+++ src/components/InfoModal/InfoModal.css
@@ -1,0 +1,39 @@
+/* Styling the modal backdrop */
+.modal-backdrop {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.5); /* Semi-transparent background */
+}
+
+/* Styling the modal content */
+.modal-content {
+    background-color: white;
+    padding: 20px;
+    border-radius: 8px;
+    position: absolute; /* Position relative to the top of the window */
+    top: 10px; /* Distance from the top of the screen */
+    left: 50%; /* Center horizontally */
+    transform: translateX(-50%); /* Ensure perfect horizontal centering */
+    max-width: 600px;
+    width: 100%;
+}
+
+/* Styling the close button (X) */
+.close-modal {
+    position: absolute;
+    top: 10px;
+    right: 15px;
+    font-size: 24px;
+    font-weight: bold;
+    color: black;
+    cursor: pointer;
+    background: transparent;
+    border: none;
+}
+
+.close-modal:hover {
+    color: red; /* Change color on hover */
+}
